Add files via upload
The input file is 'Lake_P_experiment . . . csv' and the output files for Daphnia pulicaria present (+Daphnia) or absent (-Daphnia) are 'WQfromSoilP . . . csv'. The R scripts are the Yahara water quality model. Step 1 prepares the input file for the model. Step 3 runs the model and prints the output .csv files. For this application of the model, Step 2 is not needed.